abstract more stuff to gfa_utils; empty paths
